# docs/openapi.yaml
openapi: 3.0.1
info:
  title: Bank Card & Transaction API
  description: "API для управления пользователями, банковскими картами и проведени\
    я транзакций."
  version: v1.1.0
servers:
- url: http://localhost:8080
  description: Generated server url
security:
- bearerAuth: []
tags:
- name: Аутентификация
  description: API для получения JWT-токена
- name: Управление пользователями (Admin)
  description: "API для создания, получения и удаления пользователей. Доступно тол\
    ько администраторам."
paths:
  /api/v1/auth/login:
    post:
      tags:
      - Аутентификация
      summary: Аутентификация и получение JWT-токена
      description: Обменивает имя пользователя и пароль на JWT-токен для последующих
        запросов к защищенным эндпоинтам.
      operationId: login
      requestBody:
        description: Учетные данные пользователя для входа
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRequest"
        required: true
      responses:
        "400":
          description: Некорректный формат запроса (Bad Request)
        "200":
          description: "Успешная аутентификация, токен получен"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Неверные учетные данные (Unauthorized)
  /api/v1/admin/users:
    get:
      tags:
      - Управление пользователями (Admin)
      summary: Получить список всех пользователей
      description: Возвращает полный список всех зарегистрированных пользователей.
      operationId: getAllUsers
      responses:
        "200":
          description: Успешно
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserDto"
        "401":
          description: Не авторизован
        "403":
          description: Доступ запрещен (необходима роль ADMIN)
      security:
      - bearerAuth: []
    post:
      tags:
      - Управление пользователями (Admin)
      summary: Создать нового пользователя
      operationId: createUser
      requestBody:
        description: Данные нового пользователя
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserDto"
        required: true
      responses:
        "400":
          description: Некорректные данные запроса
        "201":
          description: Пользователь успешно создан
        "401":
          description: Не авторизован
        "403":
          description: Доступ запрещен
      security:
      - bearerAuth: []
  /api/v1/admin/users/{id}:
    get:
      tags:
      - Управление пользователями (Admin)
      summary: Получить пользователя по ID
      operationId: getUserById
      parameters:
      - name: id
        in: path
        description: "ID пользователя, которого нужно найти"
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      responses:
        "200":
          description: Пользователь найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDto"
        "401":
          description: Не авторизован
        "404":
          description: Пользователь не найден
        "403":
          description: Доступ запрещен
      security:
      - bearerAuth: []
    delete:
      tags:
      - Управление пользователями (Admin)
      summary: Удалить пользователя по ID
      operationId: deleteUser
      parameters:
      - name: id
        in: path
        description: ID пользователя для удаления
        required: true
        schema:
          type: integer
          format: int64
        example: 2
      responses:
        "204":
          description: Пользователь успешно удален
        "401":
          description: Не авторизован
        "403":
          description: Доступ запрещен
      security:
      - bearerAuth: []
components:
  schemas:
    AuthRequest:
      required:
      - password
      - username
      type: object
      properties:
        username:
          maxLength: 255
          minLength: 2
          type: string
          description: Имя пользователя
          example: admin
        password:
          maxLength: 255
          minLength: 2
          type: string
          description: Пароль
          example: password
      description: Запрос на аутентификацию
    AuthResponse:
      type: object
      properties:
        token:
          type: string
          description: Сгенерированный JWT-токен
      description: Ответ с JWT-токеном
    CreateUserDto:
      required:
      - password
      - role
      - username
      type: object
      properties:
        id:
          type: integer
          description: Уникальный ID (игнорируется при создании)
          format: int64
          readOnly: true
        username:
          maxLength: 255
          minLength: 2
          type: string
          description: Имя пользователя (логин)
          example: new_user
        password:
          maxLength: 255
          minLength: 2
          type: string
          description: Пароль пользователя
          writeOnly: true
          example: Str0ngP@ssw0rd!
        role:
          type: string
          description: "Роль пользователя. Допустимые значения: ROLE_USER, ROLE_ADMIN"
          example: ROLE_USER
      description: DTO для создания нового пользователя
    UserDto:
      type: object
      properties:
        id:
          type: integer
          description: Уникальный идентификатор пользователя
          format: int64
          readOnly: true
          example: 101
        username:
          type: string
          description: Имя пользователя (логин)
          readOnly: true
          example: existing_user
        role:
          type: string
          description: Роль пользователя
          readOnly: true
          example: ROLE_ADMIN
      description: DTO с публичной информацией о пользователе (для ответа)
  securitySchemes:
    bearerAuth:
      type: http
      name: bearerAuth
      scheme: bearer
      bearerFormat: JWT